Refac longrunning & getextpath to until/host, fix portal link, markdown link helper.
