+++ src/commands/style/networkconnectivity/networkConnectivity.html
@@ -64,7 +64,7 @@
                                 <tr>
                                     <td class="table-key">{{{this.[2]}}}</td>
                                     <td class="table-value">
-                                        <p>{{#generatehtml this.[3]}}{{/generatehtml}}</p>
+                                        <p>{{#markdownHelper this.[3]}}{{/markdownHelper}}</p>
                                     </td>
                                 </tr>
                                 {{/each}}
@@ -88,7 +88,7 @@
                                     <tr>
                                         <td class="table-key">{{{this.[2]}}}</td>
                                         <td class="table-value">
-                                            <p>{{#generatehtml this.[3]}}{{/generatehtml}}</p>
+                                            <p>{{#markdownHelper this.[3]}}{{/markdownHelper}}</p>
                                         </td>
                                     </tr>
                                     
@@ -113,7 +113,7 @@
                                     <tr>
                                         <td class="table-key">{{{this.[2]}}}</td>
                                         <td class="table-value">
-                                            <p>{{#generatehtml this.[3]}}{{/generatehtml}}</p>
+                                            <p>{{#markdownHelper this.[3]}}{{/markdownHelper}}</p>
                                         </td>
                                     </tr>
                                     
@@ -139,7 +139,7 @@
                                     <tr>
                                         <td class="table-key">{{{this.[2]}}}</td>
                                         <td class="table-value">
-                                            <p>{{#generatehtml this.[3]}}{{/generatehtml}}</p>
+                                            <p>{{#markdownHelper this.[3]}}{{/markdownHelper}}</p>
                                         </td>
                                     </tr>
